python-sdk-msi-samples/python-sdk-msi-user-assigned/requirements.txt
azure-mgmt-resource>=1.1.0
azure-identity